@@ -1,121 +1,120 @@
 {
-    "name": "wavemaker-ng-runtime-new",
-    "version": "0.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "test": "ng test",
-        "test_components": "ng t @wm/components/base --code-coverage --watch=false",
-        "test_mobile": "ng t @wm/mobile/components/basic --code-coverage --watch=false",
-        "test_variables": "ng t @wm/variables --code-coverage --watch=false",
-        "test-all": "npm run test_variables && npm run test_components",
-        "test-prod": "npm run test-all > karma-test-report.txt; node combine-coverage.js >> karma-test-report.txt",
-        "lint": "ng lint",
-        "e2e": "ng e2e",
-        "optimizecss": "gulp --gulpfile build-scripts/optimize-css.gulpfile.js",
-        "postbuild": "node build-scripts/post-build.js"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "7.2.2",
-        "@angular/common": "7.2.2",
-        "@angular/compiler": "7.2.2",
-        "@angular/core": "7.2.2",
-        "@angular/forms": "7.2.2",
-        "@angular/http": "7.2.2",
-        "@angular/platform-browser": "7.2.2",
-        "@angular/platform-browser-dynamic": "7.2.2",
-        "@angular/router": "7.2.2",
-        "@ionic-native/app-version": "4.10.0",
-        "@ionic-native/barcode-scanner": "4.10.0",
-        "@ionic-native/calendar": "4.10.0",
-        "@ionic-native/camera": "4.10.0",
-        "@ionic-native/contacts": "4.10.0",
-        "@ionic-native/core": "4.10.0",
-        "@ionic-native/device": "4.10.0",
-        "@ionic-native/file": "4.10.0",
-        "@ionic-native/file-opener": "4.10.0",
-        "@ionic-native/geolocation": "4.10.0",
-        "@ionic-native/media-capture": "4.10.0",
-        "@ionic-native/network": "4.10.0",
-        "@ionic-native/sqlite": "4.10.0",
-        "@ionic-native/vibration": "4.10.0",
-        "angular-imask": "6.0.4",
-        "angular2-websocket": "0.9.7",
-        "core-js": "2.5.4",
-        "d3": "3.5.17",
-        "fullcalendar": "3.10.0",
-        "gulp": "4.0.2",
-        "gulp-brotli": "1.2.2",
-        "gulp-clean": "0.4.0",
-        "gulp-filter": "6.0.0",
-        "gulp-gzip": "1.4.2",
-        "gulp-purify-css": "0.1.0",
-        "gulp-rename": "1.4.0",
-        "hammerjs": "2.0.8",
-        "iscroll": "5.2.0",
-        "jquery": "3.3.1",
-        "jquery-ui": "1.12.1",
-        "js-cookie": "2.2.0",
-        "lodash": "4.17.11",
-        "ng-circle-progress": "1.4.0",
-        "ngx-bootstrap": "https://github.com/wavemaker/ngx-bootstrap.git#ngx-bs@build-5.4",
-        "ngx-color-picker": "7.3.0",
-        "ngx-toastr": "9.1.1",
-        "rxjs": "6.4.0",
-        "rxjs-compat": "6.4.0",
-        "summernote": "0.8.11",
-        "@wavemaker.com/nvd3": "1.0.0",
-        "x2js": "3.2.6",
-        "zone.js": "0.8.29"
-    },
-    "devDependencies": {
-        "@ampproject/rollup-plugin-closure-compiler": "0.8.5",
-        "@angular-builders/custom-webpack": "7.4.2",
-        "@angular-devkit/build-angular": "0.13.9",
-        "@angular-devkit/build-ng-packagr": "0.13.9",
-        "@angular/cli": "7.2.3",
-        "@angular/compiler-cli": "7.2.15",
-        "@angular/language-service": "7.2.2",
-        "@compodoc/compodoc": "1.1.7",
-        "@types/jquery": "3.3.22",
-        "@types/node": "8.9.4",
-        "brotli-webpack-plugin": "1.1.0",
-        "cheerio": "1.0.0-rc.2",
-        "codelyzer": "4.5.0",
-        "compression-webpack-plugin": "2.0.0",
-        "core-js-builder": "2.5.7",
-        "fs": "0.0.1-security",
-        "fs-extra": "7.0.1",
-        "istanbul-api": "1.2.2",
-        "istanbul-reports": "1.1.4",
-        "jasmine": "^3.5.0",
-        "jasmine-core": "~2.99.1",
-        "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~4.0.0",
-        "karma-chrome-launcher": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "~2.0.1",
-        "karma-jasmine": "~1.1.2",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "karma-mocha-reporter": "^2.2.5",
-        "moment": "2.24.0",
-        "ng-packagr": "4.7.1",
-        "npm-run-all": "^4.1.5",
-        "rimraf": "2.6.3",
-        "rollup": "0.65.2",
-        "rollup-plugin-alias": "1.4.0",
-        "rollup-plugin-commonjs": "9.2.0",
-        "rollup-plugin-includepaths": "0.2.3",
-        "rollup-plugin-multi-entry": "2.1.0",
-        "rollup-plugin-node-resolve": "3.4.0",
-        "ts-node": "7.0.1",
-        "tsickle": ">=0.29.0",
-        "tslib": "1.10.0",
-        "tslint": "5.11.0",
-        "typescript": "3.2.2",
-        "util": "0.11.1",
-        "@types/jasmine": "~2.8.8",
-        "@types/jasminewd2": "~2.0.3"
-    }
+  "name": "wavemaker-ng-runtime-new",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "test_components": "ng t @wm/components/base --code-coverage --watch=false",
+    "test_mobile": "ng t @wm/mobile/components/basic --code-coverage --watch=false",
+    "test_variables": "ng t @wm/variables --code-coverage --watch=false",
+    "test-all": "npm run test_variables && npm run test_components",
+    "test-prod": "npm run test-all > karma-test-report.txt; node combine-coverage.js >> karma-test-report.txt",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "optimizecss": "gulp --gulpfile build-scripts/optimize-css.gulpfile.js",
+    "postbuild": "node build-scripts/post-build.js"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "8.2.14",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
+    "@ionic-native/app-version": "4.10.0",
+    "@ionic-native/barcode-scanner": "4.10.0",
+    "@ionic-native/calendar": "4.10.0",
+    "@ionic-native/camera": "4.10.0",
+    "@ionic-native/contacts": "4.10.0",
+    "@ionic-native/core": "4.10.0",
+    "@ionic-native/device": "4.10.0",
+    "@ionic-native/file": "4.10.0",
+    "@ionic-native/file-opener": "4.10.0",
+    "@ionic-native/geolocation": "4.10.0",
+    "@ionic-native/media-capture": "4.10.0",
+    "@ionic-native/network": "4.10.0",
+    "@ionic-native/sqlite": "4.10.0",
+    "@ionic-native/vibration": "4.10.0",
+    "angular-imask": "6.0.4",
+    "angular2-websocket": "0.9.7",
+    "core-js": "2.5.4",
+    "d3": "3.5.17",
+    "fullcalendar": "3.10.0",
+    "gulp": "4.0.2",
+    "gulp-brotli": "1.2.2",
+    "gulp-clean": "0.4.0",
+    "gulp-filter": "6.0.0",
+    "gulp-gzip": "1.4.2",
+    "gulp-purify-css": "0.1.0",
+    "gulp-rename": "1.4.0",
+    "hammerjs": "2.0.8",
+    "iscroll": "5.2.0",
+    "jquery": "3.3.1",
+    "jquery-ui": "1.12.1",
+    "js-cookie": "2.2.0",
+    "lodash": "4.17.11",
+    "ng-circle-progress": "1.4.0",
+    "ngx-bootstrap": "https://github.com/wavemaker/ngx-bootstrap.git#ngx-bs@build-5.4",
+    "ngx-color-picker": "7.3.0",
+    "ngx-toastr": "9.1.1",
+    "rxjs": "6.6.0",
+    "rxjs-compat": "6.4.0",
+    "summernote": "0.8.11",
+    "@wavemaker.com/nvd3": "1.0.0",
+    "x2js": "3.2.6",
+    "zone.js": "~0.9.1"
+  },
+  "devDependencies": {
+    "@ampproject/rollup-plugin-closure-compiler": "0.8.5",
+    "@angular-builders/custom-webpack": "7.4.2",
+    "@angular-devkit/build-angular": "~0.803.28",
+    "@angular-devkit/build-ng-packagr": "~0.803.28",
+    "@angular/cli": "8.3.28",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
+    "@compodoc/compodoc": "1.1.7",
+    "@types/jquery": "3.3.22",
+    "@types/node": "8.9.4",
+    "brotli-webpack-plugin": "1.1.0",
+    "cheerio": "1.0.0-rc.2",
+    "codelyzer": "^5.0.1",
+    "compression-webpack-plugin": "2.0.0",
+    "core-js-builder": "2.5.7",
+    "fs": "0.0.1-security",
+    "fs-extra": "7.0.1",
+    "istanbul-api": "1.2.2",
+    "istanbul-reports": "1.1.4",
+    "jasmine": "^3.5.0",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-mocha-reporter": "^2.2.5",
+    "moment": "2.24.0",
+    "ng-packagr": "^5.4.0",
+    "npm-run-all": "^4.1.5",
+    "rimraf": "2.6.3",
+    "rollup": "0.65.2",
+    "rollup-plugin-alias": "1.4.0",
+    "rollup-plugin-commonjs": "9.2.0",
+    "rollup-plugin-includepaths": "0.2.3",
+    "rollup-plugin-multi-entry": "2.1.0",
+    "rollup-plugin-node-resolve": "3.4.0",
+    "ts-node": "7.0.1",
+    "tsickle": "^0.37.0",
+    "tslib": "1.13.0",
+    "tslint": "5.11.0",
+    "typescript": "3.5.3",
+    "util": "0.11.1",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3"
+  }
 }
