@@ -183,7 +183,7 @@ const markup = `<form wmForm role="" #form_1 ngNativeValidate
 })
 
 class FormWrapperComponent {
-    @ViewChild(FormComponent)
+    @ViewChild(FormComponent, /* TODO: add static flag */ {static: true})
     wmComponent: FormComponent;
     public testdata: any = [{ name: 'Ram', age: 24, dob: '2019-12-02' }, { name: 'Sita', age: 22, dob: '2019-12-02' }];
 
