@@ -37,8 +37,8 @@ export class CurrencyComponent extends NumberLocale {
     public tabindex: any;
     public shortcutkey: string;
 
-    @ViewChild(NgModel) ngModel: NgModel;
-    @ViewChild('input', {read: ElementRef}) inputEl: ElementRef;
+    @ViewChild(NgModel, /* TODO: add static flag */ {static: false}) ngModel: NgModel;
+    @ViewChild('input', /* TODO: add static flag */ { static: false, read: ElementRef }) inputEl: ElementRef;
 
     constructor(inj: Injector, i18nService: AbstractI18nService, decimalPipe: DecimalPipe) {
         super(inj, WIDGET_CONFIG, i18nService, decimalPipe);
