@@ -36,8 +36,8 @@ export class InputCalendarComponent extends BaseInput {
     public autofocus: boolean;
     public autocomplete: any;
 
-    @ViewChild('input') inputEl: ElementRef;
-    @ViewChild(NgModel) ngModel: NgModel;
+    @ViewChild('input', /* TODO: add static flag */ {static: false}) inputEl: ElementRef;
+    @ViewChild(NgModel, /* TODO: add static flag */ {static: false}) ngModel: NgModel;
 
     constructor(inj: Injector) {
         super(inj, WIDGET_CONFIG);
