@@ -1,11 +1,60 @@
 export * from './components.module';
 export * from './utils/data-utils';
-export * from './utils/form-utils';
-export * from './utils/live-utils';
-export * from './utils/widget-utils';
-export * from './widgets/common/base/base.component';
+export {
+    getOrderedDataset,
+    transformDataWithKeys,
+    extractDataAsArray,
+    convertDataToObject,
+    transformFormData,
+    getUniqObjsByDataField,
+    setItemByCompare,
+    groupData,
+    filterDate,
+    toggleAllHeaders,
+    handleHeaderClick,
+    configureDnD,
+    DataSetItem    
+} from './utils/form-utils';
+export {
+    EDIT_MODE,
+    setHeaderConfig,
+    setHeaderConfigForTable,
+    getRowOperationsColumn,
+    getFieldLayoutConfig,
+    getDefaultViewModeWidget,
+    parseValueByType,
+    getFieldTypeWidgetTypesMap,
+    getDataTableFilterWidget,
+    getEditModeWidget,
+    getDefaultValue    
+} from './utils/live-utils';
+export {
+    getObjValueByKey,
+    getEvaluatedData,
+    isActiveNavItem,
+    getOrderByExpr,
+    isDataSetWidget,
+    getImageUrl,
+    getBackGroundImageUrl,
+    provideAs,
+    provideAsWidgetRef,
+    provideAsDialogRef,
+    NAVIGATION_TYPE,
+    AUTOCLOSE_TYPE,
+    getWatchIdentifier,
+    getMatchModeTypesMap,
+    getMatchModeMsgs,
+    getConditionalClasses,
+    prepareFieldDefs    
+} from './utils/widget-utils';
+export {
+    BaseComponent
+} from './widgets/common/base/base.component';
 export * from './widgets/common/base/base-container.component';
-export * from './widgets/common/base/dataset-aware-nav.component';
+export {
+    DatasetAwareNavComponent,
+    NavNode
+} from './widgets/common/base/dataset-aware-nav.component';
 export * from './widgets/common/base/partial-container.directive';
 export * from './widgets/common/base/base-field-validations';
 export * from './widgets/common/item-template/item-template.directive';
@@ -15,11 +64,37 @@ export * from './widgets/common/pull-to-refresh/pull-to-refresh';
 export * from './widgets/common/partial/partial.directive';
 export * from './widgets/common/redraw/redrawable.directive';
 export * from './widgets/framework/constants';
-export * from './widgets/framework/deviceview';
+export {
+    updateDeviceView
+} from './widgets/framework/deviceview';
 export * from './widgets/framework/styler';
 export * from './widgets/framework/types';
-export * from './widgets/framework/widget-props';
+export {
+    getWidgetPropsByType,
+    register,
+    registerFormWidget,
+    PROP_TYPE,
+    PROP_STRING,
+    PROP_NUMBER,
+    PROP_BOOLEAN,
+    PROP_ANY    
+} from './widgets/framework/widget-props';
 export * from './widgets/common/dialog/dialog.service';
-export * from './pipes/custom-pipes';
+export {
+    ToDatePipe,
+    ToNumberPipe,
+    ToCurrencyPipe,
+    PrefixPipe,
+    SuffixPipe,
+    CustomPipe,
+    TimeFromNowPipe,
+    NumberToStringPipe,
+    StringToNumberPipe,
+    FilterPipe,
+    FileSizePipe,
+    FileIconClassPipe,
+    StateClassPipe,
+    FileExtensionFromMimePipe    
+} from './pipes/custom-pipes';
 export * from './pipes/trust-as.pipe';
 export * from './pipes/image.pipe';
