@@ -1,4 +1,6 @@
 export * from './epoch.module';
 export * from './date/date.component';
 export * from './date-time/date-time.component';
-export * from './time/time.component';
+export {
+    TimeComponent
+} from './time/time.component';
